Relocate various constructs into their own files.
